// uis/WebApp/src/app/shared/models/cart-checkout-request.ts
export interface CartCheckoutRequest {
  username: string;
  firstName: string;
  lastName: string;
  emailAddress: string;
  addressLine: string;
  country: string;
  state: string;
  zipCode: string;
  cardName: string;
  cardNumber: string;
  expiration: string;
  cvv: string;
  paymentMethod: number;
}